% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_note.R
\name{call_note}
\alias{call_note}
\title{Call note}
\usage{
call_note(ini_date, mat_date, fix_index_date, fix_spot_date, ini_date_avg,
  fin_date_avg, fx_spot, index_spot, index_ini, a, b, nom_pay, nom_prot,
  bond_cc_curve, dom_cc_curve, for_cc_curve, vol_matrix, base, coupon = 0,
  credit_spread = 0, fwd_points, haz_curve = NULL, quote_delta = TRUE,
  quote_bfrr = TRUE, delta_type = 1, rate_type = "nominal", denom = 100,
  pay_curr = "USD", nom_curr = "COP", index_curr = "USD",
  round_digits = 3, quanto = 1, quanto_id = NULL, quanto_spot = NULL,
  vol_matrix_quanto = NULL, quanto_cc_curve = NULL, rho = 0,
  prev_vol = NULL, adj_strike = 0, payoff_type = 2, index_series = NULL,
  float_strike = FALSE, geometric = FALSE, settDays, vol_factor = 1,
  M = 10000, delta_vol_max = NA, drift_cero = FALSE, rec = 0.4)
}
\arguments{
\item{ini_date}{initial date}
}
\value{
valuation
}
\description{
Valoracion de call note
}
